made changes

+++ controllers/blogs.js
@@ -1,10 +1,13 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) =>{
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user')
     response.json(blogs)
 })
+
 
 blogsRouter.post('/', async (request, response, next) =>{
     const body = request.body
@@ -13,12 +16,32 @@ blogsRouter.post('/', async (request, response, next) =>{
         return response.status(400).json({"error": 'Either url or title is missing'})
     }
 
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!decodedToken.id){
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+
+    const user = await User.findById(decodedToken.id)
+    // console.log(user)
+
     if(body.likes === undefined){
         body.likes = 0
     }
-    const blog = new Blog(body)
+    const blogObject = {
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes,
+        user: user._id
+    }
+    // console.log(blog)
+    const blog = new Blog(blogObject)
     try{
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
         response.status(201).json(savedBlog)
     }
     catch(exception){
@@ -27,6 +50,19 @@ blogsRouter.post('/', async (request, response, next) =>{
 })
 
 blogsRouter.delete('/:id', async (request, response, next) =>{
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!decodedToken.id){
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if(!(blog.user.toString() === decodedToken.id.toString())){
+        return response.status(401).json({error: 'invalid user'})
+    }
+
+    
     try{
         await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
